Support more than 64 Device without sourcecode modification
to avoid, that the source code has to be modified, when more than
64 devices shall be supported, the corresponding defines
(COM_MAXDEVICES) can be optionally set in the makefile now.

This is especially important on WSL (Windows Subsystem for Linux),
which installs 192 serial devices, no matter if there is a real
hardware counterpart on the windows side.